Peer_keys recipe: handling of etc/peer_keys
The user "Gerrit Code Review" can login with the SSH key specified in
etc/peer_keys. This is a foundation for automatic execution of Gerrit
ssh commands through Chef.

Note that etc/peer_keys only contains the second part of an
authorized_keys file (but not the ssh-(d|rsa) and the comment).